refactor: chagne in the name of HomePage's components
