+++ frontend/src/pages/home/components/footer/HomeFooter.tsx
@@ -1,11 +1,11 @@
 import React from "react";
 
-import { FooterContainer } from "./FooterStyles";
+import { FooterContainer } from "./HomeFooterStyles";
 
 import { ReactComponent as MediumIcon } from "../../../../assets/medium_icon.svg";
 import { ReactComponent as LinkedinAndInstagramIcons } from "../../../../assets/linkedin_and_insta_icons.svg";
 
-const Footer = () => {
+const HomeFooter = () => {
   return (
     <FooterContainer>
       <a href='https://medium.com/@bitwisetechnology' target="_blank"><MediumIcon /></a> 
@@ -17,4 +17,4 @@ const Footer = () => {
   );
 };
 
-export default Footer;
+export default HomeFooter;
