+++ frontend/src/pages/home/components/footer/HomeFooterStyles.ts
