+++ frontend/src/pages/home/components/header/HomeHeader.tsx
@@ -2,9 +2,9 @@ import React from 'react'
 
 import {ReactComponent as BitwiseLogo} from '../../../../assets/bitwise_logo.svg';
 
-import {StyledHeader} from './HeaderStyles';
+import {StyledHeader} from './HomeHeaderStyles';
 
-const Header = () => {
+const HomeHeader = () => {
     return (
         <StyledHeader>
             <BitwiseLogo />
@@ -12,4 +12,4 @@ const Header = () => {
     )
 }
 
-export default Header
+export default HomeHeader
