+++ frontend/src/pages/home/components/header/HomeHeaderStyles.ts
