+++ frontend/src/pages/home/components/search_container/HomeSearchContainerStyles.ts
